+++ src/parser/inline/autolink/hashtag.test.ts
@@ -10,8 +10,6 @@ describe('Unit: parser/inline/autolink/hashtag', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
       assert.deepStrictEqual(inspect(parser('#')), [['#'], '']);
       assert.deepStrictEqual(inspect(parser('# ')), [['#'], ' ']);
-      assert.deepStrictEqual(inspect(parser('#1a')), [['#1a'], '']);
-      assert.deepStrictEqual(inspect(parser('#1あ')), [['#1あ'], '']);
       assert.deepStrictEqual(inspect(parser('#a#')), [['#a#'], '']);
       assert.deepStrictEqual(inspect(parser('#a#b')), [['#a#b'], '']);
       assert.deepStrictEqual(inspect(parser('#a#b#c')), [['#a#b#c'], '']);
@@ -44,6 +42,8 @@ describe('Unit: parser/inline/autolink/hashtag', () => {
       assert.deepStrictEqual(inspect(parser('#a\\ ')), [['<a class="hashtag" href="/hashtag/a" rel="noopener">#a</a>'], '\\ ']);
       assert.deepStrictEqual(inspect(parser('#a\\\n')), [['<a class="hashtag" href="/hashtag/a" rel="noopener">#a</a>'], '\\\n']);
       assert.deepStrictEqual(inspect(parser('#あ')), [['<a class="hashtag" href="/hashtag/あ" rel="noopener">#あ</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#1a')), [['<a class="hashtag" href="/hashtag/1a" rel="noopener">#1a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#1あ')), [['<a class="hashtag" href="/hashtag/1あ" rel="noopener">#1あ</a>'], '']);
     });
 
   });