+++ src/parser/inline/comment.ts
@@ -3,7 +3,7 @@ import { validate, creator, match } from '../../combinator';
 import { html } from 'typed-dom';
 
 export const comment: CommentParser = creator(validate('<#', match(
-  /^<(#+)\s+(\S+(?:\s+(?!\1)\S+)*)(\s+\1>)?/,
+  /^<(#+)\s+(\S+(?:\s+(?!\1>)\S+)*)(\s+\1>)?/,
   ([whole, , title, closer]) => (rest, { resources }) =>
     closer
       ? [[html('sup', { class: 'comment', title })], rest]